Addresses review comments
